Add disable_ddl_transaction! to the migration template
Most of the data migrations don't want to be wrapped in a transaction, the changes on each record normally should be applied immediately. It is easy to forget about the fact that every migration is wrapped within a transaction. And not it is simple to drop this statement if not needed in a particular case.