Reset column information before running migrations
Writing to a table in one migration, amending the columns in another migration then trying to write to the same table blows up when all the migrations are in the same run. Resetting the column information on all models between migrations doesn't add much time at all, but avoids that being an issue.